requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
mysql-connector-python==9.1.0
python-dotenv==1.2.1
